*: I am so sorry
Add initial implementation of the radial library. This is such a bad
idea and I've gone through so much pain that I don't understand why I
finished writing this. It's a bad idea to begin with.

In other news, I want to know how the hell the Mac OS kernel allows you
to create mappings to the same physical memory with the mappings not
being at page boundaries (AFAIK this is a restriction of the TLB in the
MMU, not of Linux).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>